+++ Faker.Tests/LocalizationTests.cs
@@ -1,14 +1,10 @@
 ﻿using NUnit.Framework;
-using RimuTec.Faker.Helper;
-using System;
-using System.Collections.Generic;
 using System.Linq;
-using YamlDotNet.RepresentationModel;
 
 namespace RimuTec.Faker.Tests
 {
    [TestFixture]
-   public class LocalizationTests
+   public class LocalizationTests : FixtureBase
    {
       [TearDown]
       public void TearDown()
@@ -87,39 +83,5 @@ namespace RimuTec.Faker.Tests
          Assert.IsTrue(citySuffixes.Contains(citySuffix));
          Assert.Greater(citySuffixes.Count(), 0);
       }
-
-      private static List<string> Fetch(string locator)
-      {
-         var key = Config.Locale;
-         if(Config.Locale == "en")
-         {
-            key = $"{Config.Locale}.{locator.Split('.')[0]}";
-         }
-         if (_dictionary.ContainsKey(key))
-         {
-            var yamlNode = _dictionary[key];
-            var fakerNode = yamlNode;
-            var locatorParts = locator.Split('.');
-            return Fetch(yamlNode[locatorParts[0]], locatorParts.Skip(1).ToArray());
-         }
-         throw new Exception($"Entry for locale {Config.Locale} not found.");
-      }
-
-      private static List<string> Fetch(YamlNode yamlNode, string[] locatorParts)
-      {
-         if (locatorParts.Length > 0)
-         {
-            return Fetch(yamlNode[locatorParts[0]], locatorParts.Skip(1).ToArray());
-         }
-         if (yamlNode is YamlSequenceNode sequenceNode)
-         {
-            IEnumerable<string> enumerable = sequenceNode.Children.Select(c => c.ToString());
-            var arr = enumerable.ToArray();
-            return enumerable.ToList();
-         }
-         return new List<string>();
-      }
-
-      private static Dictionary<string, YamlNode> _dictionary => YamlLoader._dictionary;
    }
 }